* Merge branches/tvo3: rts
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1889 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Game/StartScripts/AirScript.h
@@ -17,7 +17,7 @@ public:
 
 	std::deque<int> planes;
 	int curPlane;
-	double lastUpdateTime;
+	float lastUpdateTime;
 	float3 tcp;
 	float3 tcf;
 	float3 oldUp[32];
